requirements.txt
PyMuPDF==1.23.8
Pillow==10.2.0
gpiozero==2.0.1
spidev==3.6
PyYAML==6.0.1
